cleanup (#37)
